Moved pushgateway to metrics/common and imports updated
